Revert "removed interleave-reads, split-paired-reads, and extract-paired-reads functions from 'test_scripts.py' and put them in new file 'test_read_handling.py'"
This reverts commit a8ae0c42dd173aa0f4e8011eca5fffbedac62041.
